Reorganize notification packages
