+++ vector/src/main/java/im/vector/app/features/settings/notifications/other/VectorSettingsOtherNotificationPreferenceFragment.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 New Vector Ltd
+ * Copyright (c) 2023 New Vector Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package im.vector.app.features.settings.notifications
+package im.vector.app.features.settings.notifications.other
 
 import im.vector.app.R
 import im.vector.app.core.preference.VectorPreferenceCategory
+import im.vector.app.features.settings.notifications.VectorSettingsPushRuleNotificationPreferenceFragment
 import org.matrix.android.sdk.api.session.pushrules.RuleIds
 
 class VectorSettingsOtherNotificationPreferenceFragment :
